# syntax/parse interpreter