--- a/racket/collects/syntax/parse/experimental/reflect.rkt
+++ b/racket/collects/syntax/parse/experimental/reflect.rkt
@@ -2,7 +2,7 @@
 (require (for-syntax racket/base
                      racket/lazy-require
                      racket/syntax
-                     (submod syntax/parse/private/residual ct)) ;; keep abs.path
+                     (submod "../private/residual.rkt" ct))
          racket/contract/base
          racket/contract/combinator
          "../private/minimatch.rkt"