--- a/racket/collects/syntax/parse/private/lib.rkt
+++ b/racket/collects/syntax/parse/private/lib.rkt
@@ -2,7 +2,7 @@
 (require "sc.rkt"
          "keywords.rkt"
          (only-in "residual.rkt" state-cons!)
-         (for-syntax syntax/parse/private/residual-ct)
+         (for-syntax (submod syntax/parse/private/residual ct))
          (for-syntax racket/base))
 
 (provide identifier