--- a/racket/collects/syntax/parse/private/litconv.rkt
+++ b/racket/collects/syntax/parse/private/litconv.rkt
@@ -5,7 +5,7 @@
                      "lib.rkt"
                      "kws.rkt"
                      racket/syntax)
-         syntax/parse/private/residual-ct ;; keep abs. path
+         (submod syntax/parse/private/residual ct) ;; keep abs. path
          syntax/parse/private/residual)   ;; keep abs. path
 (begin-for-syntax
  (lazy-require
@@ -188,6 +188,19 @@
                                                   (else ""))
                                                 " relative to the enclosing module")
                                         (quote-syntax #,stx) x))))))))]))
+
+;; (begin-for-syntax/once expr/phase1 ...)
+;; evaluates in pass 2 of module/intdefs expansion
+(define-syntax (begin-for-syntax/once stx)
+  (syntax-case stx ()
+    [(bfs/o e ...)
+     (cond [(list? (syntax-local-context))
+            #`(define-values ()
+                (begin (begin-for-syntax/once e ...)
+                       (values)))]
+           [else
+            #'(let-syntax ([m (lambda _ (begin e ...) #'(void))])
+                (m))])]))
 
 #|
 NOTES ON PHASES AND BINDINGS