--- a/racket/collects/syntax/parse/private/opt.rkt
+++ b/racket/collects/syntax/parse/private/opt.rkt
@@ -1,7 +1,7 @@
 #lang racket/base
 (require racket/syntax
          racket/pretty
-         syntax/parse/private/residual-ct ;; keep abs. path
+         (submod syntax/parse/private/residual ct) ;; keep abs. path
          "minimatch.rkt"
          "rep-patterns.rkt"
          "kws.rkt")