--- a/racket/collects/syntax/parse/private/parse.rkt
+++ b/racket/collects/syntax/parse/private/parse.rkt
@@ -13,8 +13,32 @@
                      "txlift.rkt")
          "keywords.rkt"
          syntax/parse/private/residual ;; keep abs. path
-         "runtime.rkt"
-         "parse-classic.rkt")
+         "runtime.rkt")
+
+;; ----------------------------------------
+;; Backend selection
+
+;; Interpreter backend
+#; (require "parse-interp.rkt")
+
+;; Classic backend (also disable require of "residual-interp.rkt" in "sc.rkt")
+#; (require "parse-classic.rkt")
+
+;; Hybrid backend
+(begin
+  (require (prefix-in c: "parse-classic.rkt")
+           (prefix-in i: "parse-interp.rkt"))
+  (begin-for-syntax
+    (provide interp?)
+    (define interp? (make-parameter #t))
+    (define (codegen-clauses who context x all-defs patterns body-exprs ctx track-literals?)
+      (define (call f) (f who context x all-defs patterns body-exprs ctx track-literals?))
+      (cond [(interp?) (call i:codegen-clauses)]
+            [else (call c:codegen-clauses)]))
+    (define (codegen-rhs name formals relsattrs rhs splicing? ctx)
+      (define (call f) (f name formals relsattrs rhs splicing? ctx))
+      (cond [(interp?) (call i:codegen-rhs)]
+            [else (call c:codegen-rhs)]))))
 
 ;; ============================================================