--- a/racket/collects/syntax/parse/private/parse.rkt
+++ b/racket/collects/syntax/parse/private/parse.rkt
@@ -12,7 +12,7 @@
                      "const-expr.rkt"
                      "txlift.rkt")
          "keywords.rkt"
-         syntax/parse/private/residual ;; keep abs. path
+         "residual.rkt"
          "runtime.rkt")
 
 ;; ----------------------------------------