--- a/racket/collects/syntax/parse/private/pattern-expander.rkt
+++ b/racket/collects/syntax/parse/private/pattern-expander.rkt
@@ -1,5 +1,5 @@
 #lang racket/base
-(require "residual-ct.rkt")
+(require (submod "residual.rkt" ct))
 
 (provide pattern-expander)