--- a/racket/collects/syntax/parse/private/rep-attrs.rkt
+++ b/racket/collects/syntax/parse/private/rep-attrs.rkt
@@ -1,5 +1,5 @@
 #lang racket/base
-(require syntax/parse/private/residual-ct ;; keep abs. path
+(require (submod syntax/parse/private/residual ct) ;; keep abs. path
          racket/contract/base
          syntax/private/id-table
          racket/syntax)