--- a/racket/collects/syntax/parse/private/rep-patterns.rkt
+++ b/racket/collects/syntax/parse/private/rep-patterns.rkt
@@ -1,5 +1,5 @@
 #lang racket/base
-(require (submod syntax/parse/private/residual ct) ;; keep abs. path
+(require (submod "residual.rkt" ct)
          "rep-attrs.rkt"
          "minimatch.rkt"
          "tree-util.rkt"