This PR changes the implementation of syntax/parse to use an interpreter. Specifically:
- the environment of attribute matches is represented as a data structure and only reflected as attribute bindings around user expressions
- patterns are compiled to higher-order functions that use a backtracking monad
- especially simple patterns are represented by a run-time AST and interpreted

The benefits are reduced compiled code size and, hopefully, reduced memory use. (The implementation is arguably easier to understand, too.) The cost is that syntax-parse is somewhat slower. It would be useful to know if there are cases where the slowdown is a problem.