@samth 2023-03-14
In the past, profiling Typed Racket has indicated that it spends a lot of time in syntax-parse. So I would expect a slowdown here to be bad for (the already problematic) TR compile times.