@mfelleisen 2023-03-14
Perhaps you can get @bennn to measure the time it takes to get the GTP benchmarks through.