@bennn 2023-03-15
For typechecking, the full GTP Benchmarks are overkill. I'd just use the typed code in `gtp-checkup` (`./benchmarks/*/typed/main.rkt`):
https://github.com/bennn/gtp-checkup

There's also the `tr-performance` benchmarks by @pnwamk. (I think you'll want to comment-out the plotting code and avoid the `schml` benchmarks to save time.)
https://github.com/pnwamk/tr-performance