@mflatt 2023-03-15
Here are some measurements that maybe help put a scale on the performance changes.

I tried removing all ".zo" files in the "math" collection and then ran `racket -W "debug@GC:major error" -l math` before and after this patch. As expected, the patch makes loading (mostly expansion and typechecking) take longer, but slightly less memory is used, probably because there's less code implementing the matches:

```
          CPU Time   Peak Memory
          (seconds)     (MB)
before       67         409
after        74         406
```

Rhombus also uses `syntax-parse` extensively, and the Rhombus documentation involves a lot of examples run in a sandbox. A priori, just loading the Rhombus document seems like it could go either way, since a lighter expansion of Rhombus's implementation makes the sandbox setup faster, but then the example has to be syntax-parsed. This table also shows the total ".zo" size of Rhombus's core implementation. It ends up being slightly faster.

```
rhombus.scrbl    CPU Time   Peak Memory    rhombus/private/compiled/*.zo
                (seconds)     (MB)               (MB)
before            35.6        395                9.2
after             34.6        395                7.8
```

Here are some relevant lines from the output of`dump-memory-stats` after starting DrRacket (including code still in marshaled bytevector) to show a little more about memory use:

``` 
before
 bytevector                34,572   24,832,080  |     35,603   24,895,824
 code                     149,272  131,200,464  |    149,278  131,201,744
 total                  7,228,719  501,474,116  |  7,401,631  510,486,404
after
 bytevector                34,349   22,854,528  |     35,380   22,918,272
 code                     151,348  132,071,520  |    151,354  132,072,800
 total                  7,236,443  500,788,604  |  7,409,355  509,800,892
```

Possibly not so relevant, but are some timings using `racket -cl`, which conflates the compilation of syntax-parse itself, compilation of uses of `syntax-parse,` and running some uses of `syntax-parse`. These are faster and use much less memory, which probably reflects that there are lots of uses of `syntax-parse`, and only a few are used to expand the target program.

For `racket -cl math`:

```
          CPU Time   Peak Memory
          (seconds)     (MB)
before      122         523
after       120         481
```

For `racket -ct demo.rhm` (a Rhombus demo):

```
          CPU Time   Peak Memory
          (seconds)     (MB)
before       40.5       403
after        33.1       361
```