@samth 2023-03-15
Thanks @mflatt. My view is that a 10% slowdown for Typed Racket is fairly substantial, but I like the idea of generating less code. Are there ways that we could use this change in some places but not others, perhaps limiting the runtime cost?