@mfelleisen 2023-03-15
@samth  Do you think you think @mflatt’s 10% is a definitive conclusion? Wouldn't it make sense to check what @bennn proposed? Type checking performance may depend on particular uses. (My conjecture is that it varies a lot, and we may see much more and much less than 10%.)