@mflatt 2023-03-15
I agree that more checking is needed.

@rmculpepper Did you already experiment with making small and simple patterns compiled while larger or more complex patterns are interpreted (which is what the `syntax-case` implementation does)?