@mflatt 2023-03-15
Also, because I forgot to say so before: I'm enthusiastic about this change, assuming we can find a good balance of performance. Reducing code size has been a path to steady improvement in Racket CS. Every individual change contributes only a little, but it seems to add up, because reducing code size affects things like load time, GC time, and compile time.