@mfelleisen 2023-03-26
I don't see any `++` in the "outcome" column, and my "back of the envelope" calculation (in my head) says up to 5% slowdown (on large-ish programs, like "acquire" or "schml-".  Is 5% a good trade off for space @mflatt ?