@rmculpepper 2023-03-29
Thanks Matthew and Ben for the benchmarks and measurements.

I've tried out a few changes that lead to minor improvements in the tr-performance times, but I haven't pushed them yet. They don't close the gap, in any case, and the gap is larger than 5% on most of the tr-performance benchmarks. I'd like to look at how Typed Racket is using syntax/parse in more detail, but that won't happen in time for the April release. I have reservations about keeping two implementations around in the long term, but it sounds like Matthew's heuristic is a good solution for now.