@rfindler 2023-03-29
Having two implantations can be a great help in testing, at least. :)

Robby

On Wed, Mar 29, 2023 at 2:40 PM Ryan Culpepper ***@***.***>
wrote:

> Thanks Matthew and Ben for the benchmarks and measurements.
>
> I've tried out a few changes that lead to minor improvements in the
> tr-performance times, but I haven't pushed them yet. They don't close the
> gap, in any case, and the gap is larger than 5% on most of the
> tr-performance benchmarks. I'd like to look at how Typed Racket is using
> syntax/parse in more detail, but that won't happen in time for the April
> release. I have reservations about keeping two implementations around in
> the long term, but it sounds like Matthew's heuristic is a good solution
> for now.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/racket/racket/pull/4605#issuecomment-1489195618>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AADBNMBOJ6BB6XOEZTZAIA3W6SFZVANCNFSM6AAAAAAV2MIJ5I>
> .
> You are receiving this because you are subscribed to this thread.Message
> ID: ***@***.***>
>