@mflatt 2023-05-10
@rmculpepper We're almost halfway through the v8.10 cycle, and it would be helpful to get this change in. Should I merge with the compile-syntax-classes heuristic for now?